// src/app/models/SingleNumber.model.ts
export class SingleNumber{
  // singleNumberId: number;
  // singleNumber: number;
  // numberCombinations?: {
  //   numberCombinationId: number,
  //   tripleNumber: number,
  //   visibleTripleNumber: string
  //   quantity?: number
  // }[];
  singleNumberId: number;
  singleNumber: number;
  // single_order: number;
}