// src/app/models/GameResult.model.ts
export class GameResult{

  game_date?: string;
  end_time?: string;
  triple_number?: string;
  visible_triple_number?: string;
  single_number?: string;
  result?: any[];

  draw_id?: number;
  draw_time?: string;
  multiplexer?: number;
  double_number?: string;

  rank_name?: string;
  suit_name?: string;
  rank_initial?: string;

  andar_number?: string;
  bahar_number?: string;

  // game 1

  // "draw_id": 1,
  // "draw_time": "12:02 am",
  // "multiplexer": 1,
  // "single_number": 6,
  // "double_number": "03",
  // "triple_number": "008"

  // game 2

  // "draw_id": 721,
  // "draw_time": "12:02 am",
  // "multiplexer": null,
  // "rank_name": null,
  // "suit_name": null,
  // "rank_initial": null

  // game 3

  // "draw_id": 1441,
  // "draw_time": "12:02 am",
  // "multiplexer": null,
  // "rank_name": null,
  // "suit_name": null,
  // "rank_initial": null

  // game 4

  // "draw_id": 2161,
  // "draw_time": "12:02 am",
  // "multiplexer": null,
  // "single_number": null

  // game 5

  // "draw_id": 2881,
  // "draw_time": "12:02 am",
  // "multiplexer": 1,
  // "double_number": "00",
  // "andar_number": 0,
  // "bahar_number": 0
}