// src/app/models/AndarNumber.model.ts
export class AndarNumber{
  // singleNumberId: number;
  // singleNumber: number;
  // numberCombinations?: {
  //   numberCombinationId: number,
  //   tripleNumber: number,
  //   visibleTripleNumber: string
  //   quantity?: number
  // }[];
  andarNumberId: number;
  andarNumber: number;
  // single_order: number;
}